refactor(ATT-426): split attractap websocket gateway into per-domain handlers
Extract the ~1367-line AttractapGateway god class into focused, injected
handler/service classes, leaving a thin gateway that owns only the socket
lifecycle, ACK awaiter mechanism, and EVENT routing.

New units under websockets/handlers/:
- ResourceListService      resource-list building + broadcast
- ResourceActionGuard      shared validateResourceAction
- AttractapAuthHandler     register / authenticate
- AttractapFirmwareHandler firmware info + OTA chunking
- AttractapCrashReportHandler
- AttractapCardHandler     NFC auth + card enrollment/reset
- AttractapFormsHandler    usage-form paging + draft state
- AttractapSessionHandler  start/stop session, door lock/unlock/unlatch, flow buttons
- AttractapBillingHandler  SumUp top-up
- AttractapProjectsHandler projects-of-user

Wire protocol unchanged; gateway keeps its public surface as delegators.
Door-error spec retargeted to the session handler.
