+++ apps/api/src/attractap/websockets/websocket.gateway.ts
@@ -8,8 +8,7 @@ import {
   ConnectedSocket,
 } from '@nestjs/websockets';
 import { Server } from 'ws';
-import { existsSync, statSync, openSync, readSync, closeSync } from 'fs';
-import { join } from 'path';
+import { closeSync } from 'fs';
 import { Inject, Logger, UseInterceptors } from '@nestjs/common';
 import { WebsocketService } from './websocket.service';
 import {
@@ -17,43 +16,25 @@ import {
   AttractapEvent,
   AttractapMessage,
   AttractapEventType,
-  ReaderCrashReportPayload,
 } from './websocket.types';
 import { AttractapService } from '../attractap.service';
 import { randomBytes } from 'crypto';
-import { UsersService } from '../../users-and-auth/users/users.service';
-import { AttractapFirmwareService } from '../firmware.service';
-import { SumUpService } from '../../billing/sumup.service';
 import { Mutex } from 'async-mutex';
 import { LicenseModuleType, LicenseService } from '../../license/license.service';
-import { AttractapFirmware } from '../dtos/firmware.dto';
-import { verifyToken } from './websocket.utils';
-import { ResourceUsageService } from '../../resources/usage/resourceUsage.service';
-import { ResourceMaintenanceService } from '../../resources/maintenances/maintenance.service';
-import { ResourceIntroductionsService } from '../../resources/introductions/resouceIntroductions.service';
-import { ResourceIntroducersService } from '../../resources/introducers/resourceIntroducers.service';
-import { ResourceFlowsService } from '../../resources/flows/resource-flows.service';
-import { ResourceFlowsExecutorService } from '../../resources/flows/resource-flows-executor.service';
-import { ResourceInUseError } from '../../resources/usage/errors/resource-in-use.error';
-import { InsufficientBalanceError } from '../../billing/errors/insufficient-balance.error';
-import { FlowExecutionError } from '../../resources/flows/errors/flow-execution.error';
-import { ResourceFlowNodeType, ResourceFormAction } from '@attraccess/database-entities';
-import { ProjectsService } from '../../projects/projects.service';
-import { ResourceFormsService } from '../../resources/forms/forms.service';
-import { FormSubmissionRequestDto } from '../../resources/forms/dto/form-submission-request.dto';
-import {
-  ResourceUsageFormRequestPayload,
-  ResourceUsageFormGetFieldsPayload,
-  ResourceUsageFormFieldsPayload,
-  ResourceUsageFormSubmitPagePayload,
-  ResourceUsageFormPageResultPayload,
-  FormFieldAnswerValue,
-} from './websocket.types';
 import { MetricsService } from '../../metrics/metrics.service';
 import { MetricsToggleService } from '../../metrics/settings/metrics-toggle.service';
 import { WS_METRICS } from '../../metrics/definitions/tokens';
 import { ATTRACTAP_GATEWAY_LABEL, WsMetrics } from '../../metrics/definitions/ws.metrics';
 import { WsMetricsInterceptor } from '../../metrics/instrumentation/ws/ws.interceptor';
+import { ResourceListService } from './handlers/resource-list.service';
+import { AttractapAuthHandler } from './handlers/auth.handler';
+import { AttractapFirmwareHandler } from './handlers/firmware.handler';
+import { AttractapCrashReportHandler } from './handlers/crash-report.handler';
+import { AttractapCardHandler } from './handlers/card.handler';
+import { AttractapFormsHandler } from './handlers/forms.handler';
+import { AttractapSessionHandler } from './handlers/session.handler';
+import { AttractapBillingHandler } from './handlers/billing.handler';
+import { AttractapProjectsHandler } from './handlers/projects.handler';
 
 @WebSocketGateway({ path: '/api/attractap/websocket' })
 @UseInterceptors(WsMetricsInterceptor)
@@ -70,41 +51,8 @@ export class AttractapGateway implements OnGatewayConnection, OnGatewayDisconnec
   @Inject(AttractapService)
   private attractapService: AttractapService;
 
-  @Inject(UsersService)
-  private usersService: UsersService;
-
-  @Inject(AttractapFirmwareService)
-  private firmwareService: AttractapFirmwareService;
-
   @Inject(LicenseService)
   private licenseService: LicenseService;
-
-  @Inject(ResourceUsageService)
-  private resourceUsageService: ResourceUsageService;
-
-  @Inject(ResourceMaintenanceService)
-  private resourceMaintenanceService: ResourceMaintenanceService;
-
-  @Inject(ResourceIntroductionsService)
-  private resourceIntroductionService: ResourceIntroductionsService;
-
-  @Inject(ResourceIntroducersService)
-  private resourceIntroducersService: ResourceIntroducersService;
-
-  @Inject(ResourceFlowsService)
-  private resourceFlowsService: ResourceFlowsService;
-
-  @Inject(ResourceFlowsExecutorService)
-  private resourceFlowsExecutorService: ResourceFlowsExecutorService;
-
-  @Inject(SumUpService)
-  private sumUpService: SumUpService;
-
-  @Inject(ProjectsService)
-  private projectsService: ProjectsService;
-
-  @Inject(ResourceFormsService)
-  private resourceFormsService: ResourceFormsService;
 
   @Inject(MetricsService)
   private metricsService: MetricsService;
@@ -114,6 +62,33 @@ export class AttractapGateway implements OnGatewayConnection, OnGatewayDisconnec
 
   @Inject(MetricsToggleService)
   private metricsToggle: MetricsToggleService;
+
+  @Inject(ResourceListService)
+  private resourceListService: ResourceListService;
+
+  @Inject(AttractapAuthHandler)
+  private authHandler: AttractapAuthHandler;
+
+  @Inject(AttractapFirmwareHandler)
+  private firmwareHandler: AttractapFirmwareHandler;
+
+  @Inject(AttractapCrashReportHandler)
+  private crashReportHandler: AttractapCrashReportHandler;
+
+  @Inject(AttractapCardHandler)
+  private cardHandler: AttractapCardHandler;
+
+  @Inject(AttractapFormsHandler)
+  private formsHandler: AttractapFormsHandler;
+
+  @Inject(AttractapSessionHandler)
+  private sessionHandler: AttractapSessionHandler;
+
+  @Inject(AttractapBillingHandler)
+  private billingHandler: AttractapBillingHandler;
+
+  @Inject(AttractapProjectsHandler)
+  private projectsHandler: AttractapProjectsHandler;
 
   private readonly connectedAt = new WeakMap<object, bigint>();
 
@@ -412,62 +387,62 @@ export class AttractapGateway implements OnGatewayConnection, OnGatewayDisconnec
 
     switch (eventData.type) {
       case AttractapEventType.READER_REGISTER:
-        await this.handleReaderRegister(socket, eventData);
+        await this.authHandler.handleReaderRegister(socket, eventData);
         break;
       case AttractapEventType.READER_AUTHENTICATE:
-        await this.handleAuthentication(socket, eventData);
+        await this.authHandler.handleAuthentication(socket, eventData);
         break;
       case AttractapEventType.READER_FIRMWARE_INFO:
-        await this.handleFirmwareInfo(socket, eventData);
+        await this.firmwareHandler.handleFirmwareInfo(socket, eventData);
         break;
       case AttractapEventType.READER_CRASH_REPORT:
-        await this.handleCrashReport(socket, eventData);
+        await this.crashReportHandler.handleCrashReport(socket, eventData);
         break;
       case AttractapEventType.REQUEST_CARD_AUTHENTICATION_DATA:
-        await this.handleCardAuthenticationRequest(socket, eventData);
+        await this.cardHandler.handleCardAuthenticationRequest(socket, eventData);
         break;
       case AttractapEventType.START_RESOURCE_USAGE_SESSION:
-        await this.handleStartResourceUsageSession(socket, eventData);
+        await this.sessionHandler.handleStartResourceUsageSession(socket, eventData);
         break;
       case AttractapEventType.STOP_RESOURCE_USAGE_SESSION:
-        await this.handleStopResourceUsageSession(socket, eventData);
+        await this.sessionHandler.handleStopResourceUsageSession(socket, eventData);
         break;
       case AttractapEventType.LOCK_DOOR:
-        await this.handleLockDoor(socket, eventData);
+        await this.sessionHandler.handleLockDoor(socket, eventData);
         break;
       case AttractapEventType.UNLOCK_DOOR:
-        await this.handleUnlockDoor(socket, eventData);
+        await this.sessionHandler.handleUnlockDoor(socket, eventData);
         break;
       case AttractapEventType.UNLATCH_DOOR:
-        await this.handleUnlatchDoor(socket, eventData);
+        await this.sessionHandler.handleUnlatchDoor(socket, eventData);
         break;
       case AttractapEventType.TRIGGER_FLOW_BUTTON:
-        await this.handleTriggerFlowButton(socket, eventData);
+        await this.sessionHandler.handleTriggerFlowButton(socket, eventData);
         break;
       case AttractapEventType.BILLING_REQUEST_TOPUP:
-        await this.handleBillingRequestTopup(socket, eventData);
+        await this.billingHandler.handleBillingRequestTopup(socket, eventData);
         break;
       case AttractapEventType.FIRMWARE_REQUEST_CHUNK:
-        await this.handleFirmwareChunkRequest(socket, eventData);
+        await this.firmwareHandler.handleFirmwareChunkRequest(socket, eventData);
         break;
       case AttractapEventType.ENROLL_NEW_CARD_REQUEST_NFC_KEY:
-        await this.onEnrollNewCardRequestNFCKey(socket, eventData);
+        await this.cardHandler.onEnrollNewCardRequestNFCKey(socket, eventData);
         break;
 
       case AttractapEventType.ENROLL_NEW_CARD:
-        await this.onEnrollNewCard(socket, eventData);
+        await this.cardHandler.onEnrollNewCard(socket, eventData);
         break;
 
       case AttractapEventType.PROJECTS_OF_USER:
-        await this.handleProjectsOfUserRequest(socket, eventData);
+        await this.projectsHandler.handleProjectsOfUserRequest(socket, eventData);
         break;
 
       case AttractapEventType.RESOURCE_USAGE_FORM_GET_FIELDS:
-        await this.handleResourceUsageFormGetFields(socket, eventData);
+        await this.formsHandler.handleResourceUsageFormGetFields(socket, eventData);
         break;
 
       case AttractapEventType.RESOURCE_USAGE_FORM_SUBMIT_PAGE:
-        await this.handleResourceUsageFormSubmitPage(socket, eventData);
+        await this.formsHandler.handleResourceUsageFormSubmitPage(socket, eventData);
         break;
 
       case AttractapEventType.READER_FIRMWARE_UPDATE_REQUIRED:
@@ -493,277 +468,12 @@ export class AttractapGateway implements OnGatewayConnection, OnGatewayDisconnec
     }
   }
 
-  private async handleFirmwareChunkRequest(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    try {
-      if (!socket.state.ota || !socket.state.ota.path || !socket.state.ota.size) {
-        this.logger.error(`FIRMWARE_REQUEST_CHUNK received but no OTA context set for client ${socket.id}`);
-        return;
-      }
-
-      const { offset, length } = data.payload;
-      const total = socket.state.ota.size;
-      if (typeof offset !== 'number' || typeof length !== 'number' || offset < 0 || length <= 0 || offset >= total) {
-        this.logger.error(`Invalid chunk request from client ${socket.id}: offset=${offset} length=${length}`);
-        return;
-      }
-
-      const maxChunk = 4096; // hard cap chunk size
-      const safeLen = Math.min(length, maxChunk, total - offset);
-
-      // Lazily open file descriptor
-      if (!socket.state.ota.fd) {
-        try {
-          socket.state.ota.fd = openSync(socket.state.ota.path, 'r');
-        } catch (e) {
-          this.logger.error(`Failed to open firmware for client ${socket.id}: ${(e as Error).message}`);
-          return;
-        }
-      }
-
-      const fd = socket.state.ota.fd as number;
-      const buffer = Buffer.allocUnsafe(safeLen);
-      let bytesRead = 0;
-      try {
-        bytesRead = readSync(fd, buffer, 0, safeLen, offset);
-      } catch (e) {
-        this.logger.error(`Failed to read firmware chunk for client ${socket.id}: ${(e as Error).message}`);
-        return;
-      }
-
-      if (bytesRead > 0) {
-        socket.sendBinaryData(buffer.subarray(0, bytesRead));
-        const progressPct = Math.round(((offset + bytesRead) / total) * 100);
-        const pctBucket = Math.floor(progressPct / 10) * 10;
-        const lastLogged = socket.state.ota.lastLoggedPct ?? -1;
-        if (pctBucket > lastLogged) {
-          socket.state.ota.lastLoggedPct = pctBucket;
-          this.logger.log(
-            `Attractap firmware update progress: ${pctBucket}% (reader ${socket.readerId ?? 'unknown'})`,
-          );
-        }
-      }
-    } catch (err) {
-      this.logger.error(`handleFirmwareChunkRequest error: ${(err as Error).message}`);
-    }
-  }
-
-  private async handleFirmwareInfo(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    await this.attractapService.updateReaderFirmware(socket.readerId, data.payload);
-
-    const firmwareIsUpToDate = await this.isFirmwareLatest(data.payload);
-    if (!firmwareIsUpToDate) {
-      this.logger.debug('Firmware is not up to date, notifying client to start OTA');
-
-      try {
-        const current = data.payload as AttractapFirmware;
-        const latest = await this.firmwareService.getFirmwareDefinition(current.name, current.variant);
-        if (!latest) {
-          this.logger.warn(
-            `No firmware definition found for ${current.name}/${current.variant}; treating as up-to-date for now.`,
-          );
-          return;
-        }
-
-        const assetsDir = join(__dirname, 'assets', 'attractap-firmwares');
-        const otaFilename = latest.filenameOTA || latest.filename;
-        const firmwarePath = join(assetsDir, otaFilename);
-        if (!existsSync(firmwarePath)) {
-          this.logger.error(`OTA firmware binary not found at ${firmwarePath}`);
-          return;
-        }
-        const size = statSync(firmwarePath).size;
-        if (!size || size < 1024) {
-          this.logger.error(`OTA firmware size is suspicious (${size} bytes) for ${otaFilename}`);
-          return;
-        }
-
-        // Store OTA context for this socket
-        socket.state.ota = { path: firmwarePath, size } as AuthenticatedWebSocket['state']['ota'];
-
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.READER_FIRMWARE_UPDATE_REQUIRED, {
-            available: {
-              name: latest.name,
-              variant: latest.variant,
-              version: String(latest.version),
-              totalSize: size,
-            },
-          }),
-        );
-        this.metricsService.attractapFirmwareUpdatesTotal.inc();
-      } catch (err) {
-        this.logger.error(`Failed to prepare firmware update notice: ${(err as Error).message}`);
-      }
-      return;
-    }
-  }
-
-  private async handleCrashReport(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const payload = data.payload as ReaderCrashReportPayload;
-
-    if (!payload?.resetReason || typeof payload.resetReason !== 'string') {
-      this.logger.error(`READER_CRASH_REPORT from reader ${socket.readerId} missing resetReason; ignoring.`);
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.READER_CRASH_REPORT, { error: 'INVALID_CRASH_REPORT' }),
-      );
-      return;
-    }
-
-    try {
-      const report = await this.attractapService.createCrashReport(socket.readerId, payload);
-      this.logger.warn(
-        `Stored crash report ${report.id} for reader ${socket.readerId}: reason=${report.resetReason} ` +
-          `heapFree=${report.heapFreeBytes ?? 'n/a'} largestBlock=${report.largestFreeBlockBytes ?? 'n/a'} ` +
-          `uptimeMs=${report.uptimeBeforeResetMs ?? 'n/a'} ws=${report.wsState ?? 'n/a'} wifi=${report.wifiState ?? 'n/a'}`,
-      );
-      this.metricsService.attractapCrashReportsTotal.inc();
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.READER_CRASH_REPORT, { received: true, id: report.id }),
-      );
-    } catch (error) {
-      this.logger.error(`Failed to store crash report for reader ${socket.readerId}: ${(error as Error).message}`);
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.READER_CRASH_REPORT, { error: 'CRASH_REPORT_STORE_FAILED' }),
-      );
-    }
-  }
-
-  private async isFirmwareLatest(firmware: AttractapFirmware): Promise<boolean> {
-    const firmwareDefinition = await this.firmwareService.getFirmwareDefinition(firmware.name, firmware.variant);
-
-    if (!firmwareDefinition) {
-      return true;
-    }
-
-    return String(firmwareDefinition.version) === String(firmware.version);
-  }
-
-  private async handleReaderRegister(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    this.logger.debug('Received REGISTER event');
-    const response = await this.attractapService.createNewReader(data.payload.firmware);
-
-    this.logger.debug(
-      `Sending REGISTER response to client. Reader ID: ${response.reader.id}, Token: ${response.token}`,
-    );
-
-    await socket.sendMessage(
-      new AttractapEvent(AttractapEventType.READER_REGISTER, {
-        id: response.reader.id,
-        token: response.token,
-      }),
-    );
-  }
-
-  private async handleAuthentication(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    this.logger.debug('processing READER_AUTHENTICATE event', data);
-
-    const unauthorizedResponse = new AttractapEvent(AttractapEventType.READER_UNAUTHORIZED, {
-      message: 'PLEASE_REREGISTER',
-    });
-
-    this.logger.debug('Checking if reader exists');
-    const reader = await this.attractapService.findReaderById(data.payload.id);
-    if (!reader) {
-      this.logger.error('No reader-config found for socket, sending UNAUTHORIZED response to client');
-      return await socket.sendMessage(unauthorizedResponse);
-    }
-
-    this.logger.debug('Checking if token is valid');
-    const isValidToken = await verifyToken(data.payload.token, reader.apiTokenHash);
-    if (!isValidToken) {
-      this.logger.error('Invalid token, sending UNAUTHORIZED response to client');
-      return await socket.sendMessage(unauthorizedResponse);
-    }
-
-    socket.readerId = reader.id;
-
-    const authenticatedResponse = new AttractapEvent(AttractapEventType.READER_AUTHENTICATED, {
-      name: reader.name,
-    });
-    await socket.sendMessage(authenticatedResponse);
-
-    await this.sendResourceListToSocket(socket);
-  }
-
   public async sendResourceList(readerId: number) {
-    const sockets = Array.from(this.websocketService.sockets.values()).filter((socket) => socket.readerId === readerId);
-    if (sockets.length === 0) {
-      return;
-    }
-
-    await Promise.all(sockets.map((socket) => this.sendResourceListToSocket(socket)));
+    return this.resourceListService.sendResourceList(readerId);
   }
 
   public async sendResourceListToReadersWithResource(resourceId: number) {
-    const allSockets = Array.from(this.websocketService.sockets.values());
-    await Promise.all(allSockets.map((socket) => this.sendResourceListToSocket(socket, { resourceId })));
-  }
-
-  private async sendResourceListToSocket(
-    socket: AuthenticatedWebSocket,
-    onlyIfResourceMatches?: { resourceId?: number },
-  ) {
-    const reader = await this.attractapService.findReaderById(socket.readerId);
-    if (!reader) {
-      throw new Error(`Reader not found: ${socket.readerId}`);
-    }
-
-    const resources = reader.resources;
-
-    if (onlyIfResourceMatches?.resourceId) {
-      if (!resources.some((resource) => resource.id === onlyIfResourceMatches.resourceId)) {
-        return;
-      }
-    }
-
-    const resourcesWithUsageSession = await Promise.all(
-      resources.map(async (resource) => ({
-        ...resource,
-        activeUsageSession: await this.resourceUsageService.getActiveSession(resource.id, true),
-        isUnderMaintenance: await this.resourceMaintenanceService.hasActiveMaintenance(resource.id),
-      })),
-    );
-
-    const getFlowButtons = async (resourceId: number) => {
-      const nodes = await this.resourceFlowsService.getNodes(resourceId, ResourceFlowNodeType.INPUT_BUTTON);
-      return nodes.map((node) => ({
-        id: node.id,
-        label: node.data.label || node.id,
-      }));
-    };
-
-    const resourcesWithFlowButtons = await Promise.all(
-      resourcesWithUsageSession.map(async (resource) => ({
-        ...resource,
-        flowButtons: await getFlowButtons(resource.id),
-      })),
-    );
-
-    const resourceListResponse = new AttractapEvent(AttractapEventType.RESOURCE_LIST, {
-      readerName: reader.name,
-      ledBrightness: reader.ledBrightness,
-      resources: resourcesWithFlowButtons.map((resource) => ({
-        id: resource.id,
-        name: resource.name,
-        type: resource.type,
-        separateUnlockAndUnlatch: resource.separateUnlockAndUnlatch,
-        description: resource.description,
-        allowTakeOver: resource.allowTakeOver,
-        introducers: resource.introducers.map((introducer) => introducer.user.username),
-        isUnderMaintenance: resource.isUnderMaintenance,
-        activeUsageSession: resource.activeUsageSession
-          ? {
-            user: {
-              username: resource.activeUsageSession.user.username,
-            },
-            startTime: resource.activeUsageSession.startTime.toISOString(),
-          }
-          : null,
-        flowButtons: resource.flowButtons,
-      })),
-    });
-    this.logger.debug(`Sending resource list to socket ${socket.id}`, resourceListResponse);
-    await socket.sendMessage(resourceListResponse);
+    return this.resourceListService.sendResourceListToReadersWithResource(resourceId);
   }
 
   public async disconnectReader(readerId: number) {
@@ -776,637 +486,10 @@ export class AttractapGateway implements OnGatewayConnection, OnGatewayDisconnec
   }
 
   public async startEnrollOfNewNfcCard(data: { readerId: number; userId: number }) {
-    const reader = await this.attractapService.findReaderById(data.readerId);
-
-    if (!reader) {
-      throw new Error(`Reader not found: ${data.readerId}`);
-    }
-
-    if (!reader.firmware.capabilities.cardEnrollment) {
-      throw new Error(`Reader does not support card enrollment: ${data.readerId}`);
-    }
-
-    const user = await this.usersService.findOne({ id: data.userId });
-
-    if (!user) {
-      throw new Error(`User not found: ${data.userId}`);
-    }
-
-    const sockets = Array.from(this.websocketService.sockets.values()).filter(
-      (socket) => socket.readerId === data.readerId,
-    );
-
-    if (sockets.length === 0) {
-      throw new Error(`Reader not connected: ${data.readerId}`);
-    }
-
-    // Send to all active sockets for this reader to avoid targeting a stale/disconnecting socket
-    const tasks = sockets.map(async (socket) => {
-      socket.state.lastAuthenticatedUserId = user.id;
-      try {
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD_GET_AVAILABLE_KEY_NO, {
-            username: user.username,
-          }),
-        );
-      } catch (error) {
-        // Log and continue; other sockets may still deliver the event
-        this.logger.debug(
-          `Failed to send ENROLL_NEW_CARD_GET_AVAILABLE_KEY_NO to client ${socket.id}: ${String(error)}`,
-        );
-      }
-    });
-
-    await Promise.allSettled(tasks);
-  }
-
-  private async onEnrollNewCardRequestNFCKey(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { uid, keyNo } = data.payload as { uid: string; keyNo: number };
-
-    if (!socket.state.lastAuthenticatedUserId) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD_REQUEST_NFC_KEY, { error: 'USER_NOT_SET' }),
-      );
-      return;
-    }
-
-    if (!uid || typeof uid !== 'string' || !keyNo || typeof keyNo !== 'number') {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD_REQUEST_NFC_KEY, { error: 'INVALID_PARAMS' }),
-      );
-      return;
-    }
-
-    const existingCard = await this.attractapService.getNFCCardByUID(uid);
-    if (existingCard) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD_REQUEST_NFC_KEY, { error: 'CARD_ALREADY_ENROLLED' }),
-      );
-      return;
-    }
-
-    const key = await this.attractapService.generateNTAG424Key({
-      userId: socket.state.lastAuthenticatedUserId,
-      keyNo,
-      cardUID: uid,
-    });
-
-    const keyString = this.attractapService.uint8ArrayToHexString(key);
-
-    socket.state.enrollNewCardData = {
-      keyNo,
-      key: keyString,
-      cardUID: uid,
-    };
-    await socket.sendMessage(new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD, { key: keyString, keyNo }));
-  }
-
-  private async onEnrollNewCard(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    if (!socket.state.enrollNewCardData) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD, { error: 'ENROLL_NEW_CARD_DATA_NOT_SET' }),
-      );
-      return;
-    }
-
-    const { success } = data.payload as { success: boolean };
-    if (!success) {
-      this.logger.error('Enroll new card failed');
-      return;
-    }
-
-    const { key, keyNo, cardUID } = socket.state.enrollNewCardData;
-
-    if (!key || typeof key !== 'string' || !keyNo || typeof keyNo !== 'number') {
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD, { error: 'KEY_NOT_SET' }));
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD, { error: 'USER_NOT_FOUND' }));
-      return;
-    }
-
-    await this.attractapService.createNFCCard(user, {
-      key,
-      keyNo,
-      uid: cardUID,
-    });
-
-    socket.state.enrollNewCardData = null;
-    socket.state.lastAuthenticatedUserId = null;
-    socket.sendMessage(new AttractapEvent(AttractapEventType.ENROLL_NEW_CARD, { success: true }));
+    return this.cardHandler.startEnrollOfNewNfcCard(data);
   }
 
   public async startResetOfNfcCard(data: { readerId: number; userId: number; cardId: number }) {
-    const reader = await this.attractapService.findReaderById(data.readerId);
-
-    if (!reader) {
-      throw new Error(`Reader not found: ${data.readerId}`);
-    }
-
-    const user = await this.usersService.findOne({ id: data.userId });
-
-    if (!user) {
-      throw new Error(`User not found: ${data.userId}`);
-    }
-
-    const socket = Array.from(this.websocketService.sockets.values()).find(
-      (socket) => socket.readerId === data.readerId,
-    );
-
-    if (!socket) {
-      throw new Error(`Reader not connected: ${data.readerId}`);
-    }
-
-    const nfcCard = await this.attractapService.getNFCCardByID(data.cardId);
-
-    if (!nfcCard) {
-      throw new Error(`NFC card not found: ${data.cardId}`);
-    }
-
-    // TODO: implement this
-  }
-
-  private async handleCardAuthenticationRequest(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    this.metricsService.attractapNfcTapsTotal.inc();
-    const { uid, resourceId } = data.payload as { uid: string; resourceId: number };
-
-    if (!uid || typeof uid !== 'string') {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.CARD_AUTHENTICATION_DATA, {
-          error: 'INVALID_UID',
-        }),
-      );
-      return;
-    }
-
-    const nfcCard = await this.attractapService.getNFCCardByUID(uid);
-
-    if (!nfcCard) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.CARD_AUTHENTICATION_DATA, {
-          error: 'CARD_NOT_FOUND',
-        }),
-      );
-      return;
-    }
-
-    if (!nfcCard.isActive) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.CARD_AUTHENTICATION_DATA, {
-          error: 'CARD_NOT_ACTIVE',
-        }),
-      );
-      return;
-    }
-
-    socket.state.lastAuthenticatedUserId = nfcCard.user.id;
-
-    const hasIntroduction = await this.resourceUsageService.canControllResource(resourceId, nfcCard.user);
-    const isIntroducer = await this.resourceIntroducersService.isIntroducer(resourceId, nfcCard.user.id, true);
-
-    await socket.sendMessage(
-      new AttractapEvent(AttractapEventType.CARD_AUTHENTICATION_DATA, {
-        keyNo: nfcCard.keyNo,
-        key: nfcCard.key,
-        username: nfcCard.user.username,
-        canManageResource: nfcCard.user.systemPermissions.canManageResources,
-        hasIntroduction,
-        isIntroducer,
-      }),
-    );
-  }
-
-  private async validateResourceAction(
-    socket: AuthenticatedWebSocket,
-    resourceId: number,
-    eventType: AttractapEventType,
-  ): Promise<boolean> {
-    if (!resourceId) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'INVALID_RESOURCE_ID' }),
-      );
-      return false;
-    }
-
-    const reader = await this.attractapService.findReaderById(socket.readerId);
-    if (!reader) {
-      await socket.sendMessage(new AttractapEvent(eventType, { error: 'READER_NOT_FOUND' }));
-      return false;
-    }
-
-    const resource = reader.resources.find((resource) => resource.id === resourceId);
-    if (!resource) {
-      await socket.sendMessage(
-        new AttractapEvent(eventType, {
-          error: 'RESOURCE_NOT_ASSOCIATED_WITH_READER',
-        }),
-      );
-      return false;
-    }
-
-    const lastAuthenticatedUserId = socket.state.lastAuthenticatedUserId;
-    if (lastAuthenticatedUserId == null) {
-      await socket.sendMessage(new AttractapEvent(eventType, { error: 'USER_NOT_AUTHENTICATED' }));
-      return false;
-    }
-
-    const user = await this.usersService.findOne({ id: lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(new AttractapEvent(eventType, { error: 'USER_NOT_FOUND' }));
-      return false;
-    }
-
-    return true;
-  }
-
-  private formDraftKey(resourceId: number, action: ResourceFormAction): string {
-    return `${resourceId}:${action}`;
-  }
-
-  private getFormDraft(
-    socket: AuthenticatedWebSocket,
-    resourceId: number,
-    action: ResourceFormAction,
-  ): Record<number, FormFieldAnswerValue> {
-    return socket.state.formDrafts?.[this.formDraftKey(resourceId, action)] ?? {};
-  }
-
-  private clearFormDraft(socket: AuthenticatedWebSocket, resourceId: number, action: ResourceFormAction): void {
-    if (socket.state.formDrafts) {
-      delete socket.state.formDrafts[this.formDraftKey(resourceId, action)];
-    }
-  }
-
-  private async ensureFormsSatisfied({
-    socket,
-    resourceId,
-    action,
-  }: {
-    socket: AuthenticatedWebSocket;
-    resourceId: number;
-    action: ResourceFormAction;
-  }): Promise<FormSubmissionRequestDto[] | null> {
-    const forms = await this.resourceFormsService.getFormsForAction(resourceId, action);
-    if (!forms.length) {
-      return [];
-    }
-
-    const draft = this.getFormDraft(socket, resourceId, action);
-    const hasValue = (fieldId: number) => {
-      const value = draft[fieldId];
-      return value !== undefined && value !== null && value !== '';
-    };
-
-    const complete = forms.every((form) =>
-      form.fields.filter((field) => field.isRequired).every((field) => hasValue(field.id)),
-    );
-
-    if (complete) {
-      return forms.map((form) => ({
-        formId: form.id,
-        answers: form.fields
-          .filter((field) => draft[field.id] !== undefined)
-          .map((field) => ({ fieldId: field.id, value: draft[field.id] })),
-      }));
-    }
-
-    let resourceName: string | undefined;
-    try {
-      const reader = socket.readerId ? await this.attractapService.findReaderById(socket.readerId) : null;
-      const resource = reader?.resources?.find((item) => item.id === resourceId);
-      resourceName = resource?.name;
-    } catch (error) {
-      this.logger.debug(`Failed to resolve resource name for form request: ${(error as Error).message}`);
-    }
-
-    const payload: ResourceUsageFormRequestPayload = {
-      resourceId,
-      resourceName,
-      action,
-      forms: forms.map((form) => ({ id: form.id, name: form.name, fieldCount: form.fields.length })),
-    };
-
-    await socket.sendMessage(new AttractapEvent(AttractapEventType.RESOURCE_USAGE_FORM_REQUEST, payload));
-    return null;
-  }
-
-  private async handleResourceUsageFormGetFields(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId, action, formId, offset, limit } = data.payload as ResourceUsageFormGetFieldsPayload;
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const { totalFieldCount, fields } = await this.resourceFormsService.getFieldsWindow(
-      resourceId,
-      formId,
-      offset,
-      limit,
-    );
-
-    const draft = this.getFormDraft(socket, resourceId, action);
-    const payload: ResourceUsageFormFieldsPayload = {
-      resourceId,
-      action,
-      formId,
-      offset,
-      totalFieldCount,
-      fields: fields.map((field) => ({
-        id: field.id,
-        name: field.name,
-        description: field.description ?? null,
-        type: field.type,
-        isRequired: field.isRequired,
-        options: field.options ?? null,
-        value: draft[field.id] ?? null,
-      })),
-    };
-
-    await socket.sendMessage(new AttractapEvent(AttractapEventType.RESOURCE_USAGE_FORM_FIELDS, payload));
-  }
-
-  private async handleResourceUsageFormSubmitPage(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId, action, formId, offset, answers } = data.payload as ResourceUsageFormSubmitPagePayload;
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const { valid, errors } = await this.resourceFormsService.validatePageAnswers(resourceId, formId, answers);
-
-    if (valid) {
-      const key = this.formDraftKey(resourceId, action);
-      socket.state.formDrafts ??= {};
-      socket.state.formDrafts[key] ??= {};
-      for (const answer of answers) {
-        socket.state.formDrafts[key][answer.fieldId] = answer.value;
-      }
-    }
-
-    const payload: ResourceUsageFormPageResultPayload = { resourceId, action, formId, offset, valid, errors };
-    await socket.sendMessage(new AttractapEvent(AttractapEventType.RESOURCE_USAGE_FORM_PAGE_RESULT, payload));
-  }
-
-  private async handleStartResourceUsageSession(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId, projectId } = data.payload as {
-      resourceId: number;
-      projectId?: number;
-    };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const formSubmissions = await this.ensureFormsSatisfied({
-      socket,
-      resourceId,
-      action: ResourceFormAction.START,
-    });
-    if (formSubmissions === null) {
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'USER_NOT_FOUND' }),
-      );
-      return;
-    }
-
-    try {
-      await this.resourceUsageService.startSession(resourceId, user, { projectId, formSubmissions });
-      this.clearFormDraft(socket, resourceId, ResourceFormAction.START);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { success: true }));
-    } catch (error) {
-      if (error instanceof ResourceInUseError) {
-        setTimeout(async () => {
-          await this.sendResourceListToSocket(socket, { resourceId });
-        }, 1000);
-        return;
-      }
-      if (error instanceof InsufficientBalanceError || error?.message === 'INSUFFICIENT_BALANCE') {
-        const sumUpEnabled = await this.sumUpService.getIsEnabled();
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, {
-            error: 'INSUFFICIENT_BALANCE',
-            sumUpEnabled,
-          }),
-        );
-        return;
-      }
-      this.logger.error(`Failed to start resource usage session: ${error.message}`);
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: error.message }),
-      );
-    }
-  }
-
-  private async handleStopResourceUsageSession(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId } = data.payload as {
-      resourceId: number;
-    };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const formSubmissions = await this.ensureFormsSatisfied({
-      socket,
-      resourceId,
-      action: ResourceFormAction.END,
-    });
-    if (formSubmissions === null) {
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'USER_NOT_FOUND' }),
-      );
-      return;
-    }
-
-    try {
-      await this.resourceUsageService.endSession(resourceId, user, { formSubmissions });
-      this.clearFormDraft(socket, resourceId, ResourceFormAction.END);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.STOP_RESOURCE_USAGE_SESSION, { success: true }));
-    } catch (error) {
-      this.logger.error(`Failed to stop resource usage session: ${error.message}`);
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.STOP_RESOURCE_USAGE_SESSION, { error: error.message }),
-      );
-    }
-  }
-
-  private async handleLockDoor(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId } = data.payload as { resourceId: number };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'USER_NOT_FOUND' }),
-      );
-      return;
-    }
-
-    try {
-      await this.resourceUsageService.lockDoor(resourceId, user);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.LOCK_DOOR, { success: true }));
-    } catch (error) {
-      const errorMessage = this.extractUserFacingError(error);
-      this.logger.error(`Failed to lock door: ${errorMessage}`);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.LOCK_DOOR, { error: errorMessage }));
-    }
-  }
-
-  private async handleUnlockDoor(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId } = data.payload as { resourceId: number };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'USER_NOT_FOUND' }),
-      );
-      return;
-    }
-
-    try {
-      await this.resourceUsageService.unlockDoor(resourceId, user);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.UNLOCK_DOOR, { success: true }));
-    } catch (error) {
-      const errorMessage = this.extractUserFacingError(error);
-      this.logger.error(`Failed to unlock door: ${errorMessage}`);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.UNLOCK_DOOR, { error: errorMessage }));
-    }
-  }
-
-  private async handleUnlatchDoor(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId } = data.payload as { resourceId: number };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.START_RESOURCE_USAGE_SESSION))) {
-      return;
-    }
-
-    const user = await this.usersService.findOne({ id: socket.state.lastAuthenticatedUserId });
-    if (!user) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.START_RESOURCE_USAGE_SESSION, { error: 'USER_NOT_FOUND' }),
-      );
-      return;
-    }
-
-    try {
-      await this.resourceUsageService.unlatchDoor(resourceId, user);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.UNLATCH_DOOR, { success: true }));
-    } catch (error) {
-      const errorMessage = this.extractUserFacingError(error);
-      this.logger.error(`Failed to unlatch door: ${errorMessage}`);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.UNLATCH_DOOR, { error: errorMessage }));
-    }
-  }
-
-  private async handleTriggerFlowButton(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const { resourceId, buttonId } = data.payload as { resourceId: number; buttonId: string };
-
-    if (!(await this.validateResourceAction(socket, resourceId, AttractapEventType.TRIGGER_FLOW_BUTTON))) {
-      return;
-    }
-
-    try {
-      await this.resourceFlowsExecutorService.pressButton(resourceId, buttonId, socket.state.lastAuthenticatedUserId);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.TRIGGER_FLOW_BUTTON, { success: true }));
-    } catch (error) {
-      this.logger.error(`Failed to trigger flow button: ${error.message}`);
-      await socket.sendMessage(new AttractapEvent(AttractapEventType.TRIGGER_FLOW_BUTTON, { error: error.message }));
-    }
-  }
-
-  private async handleBillingRequestTopup(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    try {
-      const enabled = await this.sumUpService.getIsEnabled();
-      if (!enabled) {
-        this.logger.warn('BILLING_REQUEST_TOPUP received but SumUp is not enabled');
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.BILLING_REQUEST_TOPUP, { error: 'SUMUP_NOT_ENABLED' }),
-        );
-        return;
-      }
-
-      const userId = socket.state.lastAuthenticatedUserId;
-      if (!userId) {
-        this.logger.warn('BILLING_REQUEST_TOPUP received but no authenticated user is set on socket');
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.BILLING_REQUEST_TOPUP, { error: 'USER_NOT_AUTHENTICATED' }),
-        );
-        return;
-      }
-
-      const { amountCents } = data.payload as { amountCents: number };
-      if (typeof amountCents !== 'number' || !Number.isInteger(amountCents) || amountCents <= 0) {
-        this.logger.warn(`BILLING_REQUEST_TOPUP invalid amount: ${JSON.stringify(data.payload)}`);
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.BILLING_REQUEST_TOPUP, { error: 'INVALID_AMOUNT' }),
-        );
-        return;
-      }
-
-      const readers = await this.sumUpService.getReaders();
-      if (!readers || readers.length === 0) {
-        this.logger.warn('BILLING_REQUEST_TOPUP requested but no SumUp readers are linked');
-        await socket.sendMessage(
-          new AttractapEvent(AttractapEventType.BILLING_REQUEST_TOPUP, { error: 'NO_SUMUP_TERMINALS_AVAILABLE' }),
-        );
-        return;
-      }
-
-      // Prefer a paired/active reader if available, otherwise first one
-      const preferred = readers.find((r) => r.status === 'paired') || readers[0];
-      await this.sumUpService.topUpWithReader(userId, preferred.id, amountCents);
-      this.logger.debug(`Started SumUp top-up for user ${userId} on reader ${preferred.id} amount ${amountCents}`);
-    } catch (err) {
-      this.logger.error(`handleBillingRequestTopup error: ${(err as Error).message}`);
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.BILLING_REQUEST_TOPUP, {
-          error: 'SUMUP_TOPUP_FAILED',
-          details: (err as Error).message,
-        }),
-      );
-    }
-  }
-
-  private async handleProjectsOfUserRequest(socket: AuthenticatedWebSocket, data: AttractapEvent['data']) {
-    const userId = socket.state.lastAuthenticatedUserId;
-    if (!userId) {
-      await socket.sendMessage(
-        new AttractapEvent(AttractapEventType.PROJECTS_OF_USER, { error: 'USER_NOT_AUTHENTICATED' }),
-      );
-      return;
-    }
-
-    const { page = 1, limit = 10 } = data.payload as { page?: number; limit?: number };
-
-    const projects = await this.projectsService.findMany(userId, { page, limit });
-    const total = await this.projectsService.getTotalCount(userId);
-    await socket.sendMessage(new AttractapEvent(AttractapEventType.PROJECTS_OF_USER, { projects, page, limit, total }));
-  }
-
-  private extractUserFacingError(error: unknown): string {
-    if (error instanceof FlowExecutionError) {
-      return error.message.replace(/^FLOW_EXECUTION_ERROR:\s*/, '');
-    }
-    return error instanceof Error ? error.message : String(error);
+    return this.cardHandler.startResetOfNfcCard(data);
   }
 }
